renaming whole folder to rename and re-import all files
