+++ component/AlertNotice/AlertNotice.tsx
