+++ component/AlertNotice/index.tsx
