+++ component/ComponentSVG/index.tsx
