+++ component/SlinkyCursor/index.tsx
