--- components/footer/footer.module.css
@@ -1,25 +1,0 @@
-.footer {
-  bottom: 0;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  width: 100%;
-  padding: var(--size-04) var(--size-06);
-  min-height: var(--size-09);
-}
-.footer ul {
-  display: flex;
-  gap: var(--size-05);
-  padding: 0;
-  margin: 0;
-  list-style: none;
-}
-@media screen and (min-width: 1000px) {
-  .footer {
-    max-width: 1280px;
-    margin: 0 auto;
-  }
-  .footer ul {
-    flex-direction: row;
-  }
-}
