--- components/navbar/index.module.css
@@ -1,49 +1,0 @@
-.logo {
-  width: var(--size-07);
-  height: var(--size-07);
-  margin: 0;
-}
-.nav a:has(svg) {
-  display: flex;
-  align-items: center;
-  gap: var(--size-02);
-  color: var(--color-foreground);
-}
-.nav {
-  display: flex;
-  flex-direction: column;
-  gap: var(--size-02);
-  min-height: 1rem;
-  padding: var(--size-02) var(--size-06);
-  margin-bottom: var(--size-02);
-}
-.nav ul {
-  display: flex;
-  align-items: center;
-  gap: var(--size-05);
-  list-style-type: none;
-  padding: 0;
-  margin: 0;
-}
-.nav h1{
-  line-height: 1;
-  margin-top: 0.5rem;
-  font-weight: var(--font-weight-light) !important;
-} 
-.nav a.active {
-  font-weight: var(--font-weight-bold);
-  color: var(--foreground);
-}
-.nav a { 
-  font-family: var(--font-nohemi);  
-}
-@media screen and (min-width: 1000px) {
-  .nav {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    max-width: 1280px;
-    margin: 0 auto;
-  }
-}
