--- components/navbar/line-logo.tsx
@@ -1,20 +1,0 @@
-const LineLogo = () => {
-  return (
-    <svg
-      width="36"
-      height="36"
-      viewBox="0 0 36 36"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        style={{ stroke: "var(--foreground)" }}
-        d="M25 15L23.5 8M25 15L33 16M25 15L18 17M23.5 8L33 16M23.5 8H12.5M33 16L18 31M12.5 8L11 15M12.5 8L3 16M11 15L3 16M11 15L18 17M3 16L18 31M18 17V31"
-        stroke-width="1.5"
-        stroke-linejoin="round"
-      />
-    </svg>
-  );
-};
-
-export default LineLogo;
