--- components/navbar/logo.tsx
@@ -1,54 +1,0 @@
-const Logo = () => {
-  return (
-    <svg
-      width="36"
-      height="36"
-      viewBox="0 0 36 36"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M23.5 8L25 15L33 16L23.5 8Z"
-        fill="#E74B4C"
-        stroke="#1A1B1C"
-        strokeLinejoin="round"
-      />
-      <path
-        d="M23.5 8L25 15M33 16L25 15M18 31V17M3 16L11 15M12.5 8L11 15M11 15L18 17M18 17L25 15"
-        stroke="#1A1B1C"
-        strokeLinejoin="round"
-      />
-      <path
-        d="M25 15L18 17V31L33 16L25 15Z"
-        fill="#963131"
-        stroke="#1A1B1C"
-        strokeLinejoin="round"
-      />
-      <path
-        d="M11 15L12.5 8H23.5L25 15L18 17L11 15Z"
-        fill="#E73C3E"
-        stroke="#1A1B1C"
-        strokeLinejoin="round"
-      />
-      <path
-        d="M3 16L11 15L18 17V31L3 16Z"
-        fill="#D14345"
-        stroke="#1A1B1C"
-        strokeLinejoin="round"
-      />
-      <path
-        d="M11 15L12.5 8L3 16L11 15Z"
-        fill="#FF616A"
-        stroke="#1A1B1C"
-        strokeLinejoin="round"
-      />
-      <path
-        d="M33 16L23.5 8H12.5L3 16L18 31L33 16Z"
-        stroke="#1A1B1C"
-        strokeWidth="1.5"
-        strokeLinejoin="round"
-      />
-    </svg>
-  );
-};
-export default Logo;
