--- components/tiptap/CodeBlockComponent.tsx
@@ -1,26 +1,0 @@
-import './CodeBlockComponent.css'
-
-import { NodeViewContent, NodeViewWrapper } from '@tiptap/react'
-import React from 'react'
-
-const codeBlock = ({ node: { attrs: { language: defaultLanguage } }, updateAttributes, extension }) => (
-  <NodeViewWrapper className="code-block">
-    <select contentEditable={false} defaultValue={defaultLanguage} onChange={event => updateAttributes({ language: event.target.value })}>
-      <option value="null">
-        auto
-      </option>
-      <option disabled>
-        —
-      </option>
-      {extension.options.lowlight.listLanguages().map((lang, index) => (
-        <option key={index} value={lang}>
-          {lang}
-        </option>
-      ))}
-    </select>
-    <pre>
-      <NodeViewContent as="code" />
-    </pre>
-  </NodeViewWrapper>
-)
-export default codeBlock
