--- components/toggleSwitch/toggleSwitch.module.css
@@ -1,47 +1,0 @@
-.toggleSwitch {
-  display: flex;
-  position: relative;
-  max-height: 1.5rem;
-  gap: 0.25rem;
-  align-items: center;
-  line-height: 1;
-}
-.toggleSwitch input {
-  all: unset;
-  cursor: pointer;
-  width: 3rem;
-  height: 1.5rem;
-  border-radius: var(--radius-full);
-
-  border: 1px solid var(--foreground);
-  background-color: var(--background-secondary); 
-  transition: all 0.3s ease;
-}
-.toggleSwitch input.checked {
-  background-color: var(--color-accent);
-} 
-.toggleSwitch label::before {
-  position: absolute;
-  display: inline-block;
-  top: 0.2125rem;
-  left: 0.25rem;
-  content: " ";
-  height: 1rem;
-  width: 1rem;
-  background-color: var(--background);
-  border: 1px solid var(--foreground);
-  border-radius: var(--radius-full);
-  transition: all ease 0.3s;
-}
-.toggleSwitch label.checked::before {
-  left: 1.75rem;
-  background-color: var(--foreground);
-}
-.toggleSwitch input:focus-visible {
-  outline: 0.2rem solid var(--color-accent);
-  outline-offset: 0.2rem;
-}
-.toggleSwitch label {
-  margin-left: 0.25rem;
-  cursor: pointer;
-}
